dotstracks/requirements.txt
Django==3.2.5
django-cors-headers
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
PyJWT==1.7.1
gunicorn>=20.1
django-redis
psycopg2-binary
pika
Faker
redis
progress